Changed file format drastically, moved to Maven, & huge changes to classes
If someone far in the future is reading this, sorry for the really vague and weird commit. I wanted to change a lot in the code, then start committing in small pieces afterwards for changes.

The reason for this is that the original code that I uploaded in the first or second commit was some code I pulled from a tutorial, so this is technically the first commit with a lot of drastic changes that I've implemented myself, such as moving a lot of things to their own classes.

Future commits are to be pushed in small bits and to be very clear.